simplify matchplus and matchstar
`matchplus` can be simplified by only modifying `matchlength` once the
complete match is successful.  This means it doesn't have to rewind
`matchlength` as it iterates through each possible `matchpattern`.
This also means it keeps `matchlength` unmodified if it doesn't return
a match.  Because of this last part, this also means that `matchstar`
can leverage `matchplus` which reduces it to single line of code
`return matchplus(...) || matchpattern(..)`.