fix python regex test scripts
In python, when you pass an array to "call", the arguments are not to be quoted unless you are intending on passing the quotes to the underlying command you are running.  In this case the the underlying programs test_rand.c and test_rand_neg.c are not expecting their arguments to be quoted, but instead, are expecting them to be unquoted regular expression strings.  This is a common mistake because people confuse the BASH syntax for escaping characters with the strings passed to a subprocess in Python.